chore(node): further integrate and organize dir Also cleans up usings.
