+++ sn/src/node/routing/api/command.rs
@@ -6,9 +6,8 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::messaging::system::SectionAuth;
 use crate::messaging::{
-    system::{DkgFailureSigSet, KeyedSig, NodeState, SystemMsg},
+    system::{DkgFailureSigSet, KeyedSig, NodeState, SectionAuth, SystemMsg},
     DstLocation, MessageId, NodeMsgAuthority, WireMsg,
 };
 use crate::node::routing::{
