+++ sn/src/node/routing/core/bootstrap/mod.rs
@@ -13,9 +13,10 @@ pub(crate) use join::join_network;
 pub(crate) use relocate::JoiningAsRelocated;
 
 #[cfg(not(test))]
-use crate::node::routing::Error;
-use crate::node::routing::Result;
+use crate::node::Error;
+use crate::node::Result;
 use crate::prefix_map::NetworkPrefixMap;
+
 use bls::PublicKey as BlsPublicKey;
 use std::{collections::HashSet, net::SocketAddr};
 #[cfg(not(test))]
