+++ sn/src/node/routing/core/msg_handling/update_section.rs
@@ -6,8 +6,10 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::Core;
-use crate::node::routing::{error::Result, routing_api::command::Command};
+use crate::node::{
+    error::Result,
+    routing::{api::command::Command, core::Core},
+};
 use std::collections::BTreeSet;
 
 impl Core {
