+++ sn/src/node/routing/core/proposal.rs
@@ -6,11 +6,11 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::{
-    messaging::system::{Proposal as ProposalMsg, SectionAuth},
-    node::routing::{
+use crate::messaging::system::{Proposal as ProposalMsg, SectionAuth};
+use crate::node::{
+    error::Result,
+    routing::{
         dkg::SigShare,
-        error::Result,
         network_knowledge::{NodeState, SectionAuthorityProvider},
     },
 };
