+++ sn/src/prefix_map/mod.rs
@@ -18,14 +18,21 @@
 mod stats;
 
 use self::stats::NetworkStats;
+
 use crate::messaging::system::SectionAuth;
-use crate::node::routing::{Error, Result, SectionAuthUtils, SectionAuthorityProvider};
+use crate::node::{
+    routing::{SectionAuthUtils, SectionAuthorityProvider},
+    Error, Result,
+};
+
 use bls::PublicKey as BlsPublicKey;
 use dashmap::{self, mapref::multiple::RefMulti, DashMap};
 use secured_linked_list::SecuredLinkedList;
 use serde::{Deserialize, Serialize};
-use std::iter::{self, Iterator};
-use std::sync::Arc;
+use std::{
+    iter::{self, Iterator},
+    sync::Arc,
+};
 use xor_name::{Prefix, XorName};
 
 /// Container for storing information about other sections in the network.
